requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
transformers==4.44.2
torch==2.4.1
redis==5.0.8
shap==0.46.0
pydantic==2.9.2
pyyaml==6.0.2
sqlite3
smolagents==0.1.0